adds segmentation folder gan

+++ preprocess_Cohen-Kaggle-BIMCV-OCT.py
@@ -4,23 +4,30 @@ import re
 import random
 import cv2
 import numpy as np
+import telebot
 # Split data in training/testing using a 80/20 distribution
 # The distribution is centered around the patient since each person can have multiple images in different days
 # So the idea is to keep all images from the same patient all in the same set (either train or test)
+TELEBOT_TOKEN = "2058519653:AAG5Kf0Othtye8e13F5WPnBQQSdoCt47ifA"
 
+bot = telebot.TeleBot("2058519653:AAG5Kf0Othtye8e13F5WPnBQQSdoCt47ifA")
+bot.config['api_key'] = TELEBOT_TOKEN
+bot.get_me()
 
 # The dataset is composed by CXR images of pneumonia (any other, except COVID-19), COVID-19 and Normal.
-source_folders = [ "BIMCV", "OCT"] #"Cohen", "RSNA", "Actualmed", "Figure1", "KaggleCRD", "RICORD"
+source_folders = [ "Cohen", "OCT", "Actualmed", "Figure1", "KaggleCRD", "BIMCV"] #"RICORD" #RSNA
 pneumonia_folders = ["Bacteria", "Fungi", "Virus", "Pneumonia", "Lung Opacity"]
 pathogen_folders = ["Bacteria", "Fungi", "Virus", "Pneumonia", "Lung Opacity", "COVID-19", "Normal"]
-origin_folder = "2_Raw_Seg"
-dest_folder = "3_Images_Tests_Seg"
+origin_folder = "2_Raw"
+dest_folder = "3_Images"
 
 img_size = 400
 cwd = os.getcwd()
 
 clahe = cv2.createCLAHE(clipLimit = 2.0, tileGridSize = (8, 8))
 
+
+bot.send_message("-600800507", f"Iniciando separação de treino e teste")
 # Remove destination dir, if present, just for start clean
 if os.path.isdir(dest_folder):
   shutil.rmtree(dest_folder)
@@ -52,12 +59,13 @@ for folder in source_folders:
       for (root, dirs, files) in os.walk(pathogen_folder, topdown = True):
         pid_list = {}
         for file in files:
+          print(file)
           _, pid, offset = re.split("[P_]", os.path.splitext(file)[0])
 
           # If pid was not assigned to a group
           # Then random selected train/test following the desirable distribution
           if pid not in pid_list:
-            prob = 0.7
+            prob = 0.8
             target_folder = "test" if random.uniform(0, 1) > prob else "train"
             pid_list[pid] = target_folder
           else:
@@ -68,19 +76,24 @@ for folder in source_folders:
           shutil.copy2(os.path.join(root, file), os.path.join(dest_folder, target_folder, pathogen_coded))
           new_filename = "%s_%s_%s_%s" % (folder, pathogen_coded, pid, offset)
           new_filename_ext = "%s%s" % (new_filename, os.path.splitext(file)[-1])
-          os.rename(
-            os.path.join(dest_folder, target_folder, pathogen_coded, file),
-            os.path.join(dest_folder, target_folder, pathogen_coded, new_filename_ext),
-          )
+
+          new_name = os.path.join(dest_folder, target_folder, pathogen_coded, new_filename_ext)
+          
+          if not os.path.exists(new_name):
+            os.rename(
+              os.path.join(dest_folder, target_folder, pathogen_coded, file),
+              os.path.join(dest_folder, target_folder, pathogen_coded, new_filename_ext),
+            )
 
           # Well, let's already apply CLAHE to improve the CXR contrast and brightness
-          img = cv2.imread(os.path.join(dest_folder, target_folder, pathogen_coded, new_filename_ext), 0)
+          img = cv2.imread(os.path.join(dest_folder, target_folder, pathogen_coded, new_filename_ext), cv2.IMREAD_COLOR)
           os.remove(os.path.join(dest_folder, target_folder, pathogen_coded, new_filename_ext))
           #img = clahe.apply(img)
 
           # Let's also resize the images so that all of them are standardize
           # Skip the image if it is too small
-          w, h = img.shape
+          print(img.shape)
+          w, h = img.shape[:-1]
           if w < 250:
             continue
 
@@ -98,3 +111,27 @@ for folder in source_folders:
             mask_img = cv2.resize(mask_img, (img_size, img_size), interpolation = cv2.INTER_CUBIC)
             new_mask_filename = "%s_%s_%s_%s_%s.png" % (target_folder, folder, pathogen_coded, pid, offset)
             cv2.imwrite(os.path.join(dest_folder, "masks", new_mask_filename), mask_img)
+
+
+gan = False
+
+if (gan == True):
+
+  source = '/home/dell/Documentos/covid-dissert/augmentation/Results/Pneumonia_Imgs/'
+  files = os.listdir(source)
+  dest = "/home/dell/Documentos/covid-dissert/3_Images/train/Opacity/"
+  for file_name in random.sample(files, 500):
+    shutil.copy2(os.path.join(source, file_name), dest)
+  
+  source = '/home/dell/Documentos/covid-dissert/augmentation/Results/COVID-19_Imgs/'
+  files = os.listdir(source)
+  dest = "/home/dell/Documentos/covid-dissert/3_Images/train/COVID-19/"
+  for file_name in random.sample(files, 500):
+    shutil.copy2(os.path.join(source, file_name), dest)
+
+
+  source = '/home/dell/Documentos/covid-dissert/augmentation/Results/Normal_Imgs/'
+  files = os.listdir(source)
+  dest = "/home/dell/Documentos/covid-dissert/3_Images/train/Normal/"
+  for file_name in random.sample(files, 500):
+    shutil.copy2(os.path.join(source, file_name), dest)
